feat: move getCurrentOnChainDistribution
